granted: support fish shell
Due to fish being a non-POSIX shell, granted provides a separate wrapper
to be used by fish.  The NixOS package for granted allows a user to
enable the fish wrapper if they enable it via an override.

https://github.com/NixOS/nixpkgs/blob/3fd3383906eb781676838a3c51c9212a6f803cce/pkgs/by-name/gr/granted/package.nix#L79

This PR adds a fish integration and enables the withFish override if the
fish integration is enabled.